+++ RobotControl/robot/structures.h
@@ -20,7 +20,7 @@ uint8_t getByte(const T& value)
 }
 
 uint16_t toUInt16(const QByteArray& data, const int index);
-uint16_t toUInt32(const QByteArray& data, const int index);
+uint32_t toUInt32(const QByteArray& data, const int index);
 
 /**
  * @brief Liste des commandes
